rss/requirements.txt
feedparser
psycopg2-binary
beautifulsoup4
requests
